Revert "Aktualizacja biblioteki SuplaDevice"
This reverts commit 68470d2ddd96412005e7feb7fc074270968ac5fb.

+++ lib/SuplaDevice/extras/porting/linux/supla/control/cmd_relay.h
@@ -45,7 +45,7 @@ class CmdRelay : public Sensor::SensorParsed<VirtualRelay> {
  protected:
   std::string cmdOn;
   std::string cmdOff;
-  uint32_t lastReadTime = 0;
+  uint64_t lastReadTime = 0;
   bool useOfflineOnInvalidState = false;
 };
