+++ lib/SuplaDevice/src/supla/control/bistable_roller_shutter.h
@@ -39,7 +39,7 @@ class BistableRollerShutter : public RollerShutter {
   void relayDownOff() override;
 
   bool activeBiRelay = false;
-  uint32_t toggleTime = 0;
+  uint64_t toggleTime = 0;
 };
 
 }  // namespace Control
