+++ lib/SuplaDevice/src/supla/sensor/pressure.h
@@ -46,7 +46,7 @@ class Pressure : public ChannelElement {
   }
 
  protected:
-  uint32_t lastReadTime;
+  uint64_t lastReadTime;
 };
 
 };  // namespace Sensor
