+++ lib/SuplaDevice/src/supla/sensor/rain.h
@@ -46,7 +46,7 @@ class Rain : public ChannelElement {
   }
 
  protected:
-  uint32_t lastReadTime;
+  uint64_t lastReadTime;
 };
 
 };  // namespace Sensor
